+++ blocklets/ai-studio/src/pages/project/debug-view.tsx
@@ -531,6 +531,7 @@ function ChatModeForm({
           type="submit"
           variant="contained"
           sx={{
+            whiteSpace: 'nowrap',
             background: '#030712',
             color: '#fff',
             '&:hover': {
@@ -631,7 +632,13 @@ function DebugModeForm({
 
     sendMessage({
       sessionIndex: state.currentSessionIndex!,
-      message: { type: 'debug', projectId, assistantId: assistant.id, gitRef, parameters },
+      message: {
+        type: 'debug',
+        projectId,
+        assistantId: assistant.id,
+        gitRef,
+        parameters: { ...parameters, $clientTime: new Date().toISOString() },
+      },
     });
     setSession(state.currentSessionIndex!, (session) => {
       session.debugForm = { ...parameters };