+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.338 (2024-4-17)
+## 0.1.339 (2024-4-17)
 
 - fix: copy tool(#801)
 - fix: visit link in dataset discussion(#795)
@@ -9,6 +9,11 @@
 - fix: output name i18n(#792)
 - fix: add default title for publish view #811
 - fix: add new one tool when update too #809
+
+## 0.1.338 (2024-4-17)
+
+- feat: add `$clientTime` parameter
+- fix: avoid button title wrap
 
 ## 0.1.337 (2024-4-16)